Part 10 done

+++ scr/scripts/journal.js
@@ -1,6 +1,7 @@
 import API from "./data.js"
 import renderJournalEntries from "./entriesDOM.js"
 
+
 // Entry Input Object Factory
 let entryFactory = (date, concepts, content, mood) => {
     return {
@@ -19,40 +20,83 @@ document.querySelector("#buttonSubmit").addEventListener("click", () => {
     let concepts = document.querySelector("#cenceptsForm").value;
     let content = document.querySelector("#contentForm").value;
     let mood = document.querySelector("#moodForm").value;
+    let idEdit = document.querySelector("#recipeId").value;
+
+    const objeto = entryFactory(date, concepts, content, mood)
 
     // Garantee that the object has all values
     if (!!date && !!concepts && !!content && !!mood) {
 
         // POST the object in the JSON-server
-        API.post(entryFactory(date, concepts, content, mood))
+        if(!idEdit){
+        API.post(objeto)
             .then(data => {
 
-                // render the new data in the Past Entries
-                API.get()
-                    .then(data => renderJournalEntries(data))
+                // render the updated database
+                API.get().then(data => renderJournalEntries(data))
             }
             )
+        }
+
+        // Edit the object in the JSON-server
+        else {
+            // PUT
+            API.put(objeto, idEdit)
+            .then(a => {
+                //Reset the idEdit.value
+                idEdit.value = ""
+
+                // render the updated database
+                alert("Entry edited sucessfuly!")
+                API.get().then(data => renderJournalEntries(data))
+            })
+        }
     } else {
         alert("Please compelte the entry in order to submit")
     }
 })
 
-// Detele entry
+// Detele / Edit  entry
 const entryLoga = document.getElementById("entryLog")
 entryLoga.addEventListener("click", event => {
-    let entryID = event.target.id.split("--")[2]
+    let idArray = event.target.id.split("--")
+    let entryID = idArray[2]
 
-    // Pop Up Confirm
-    if (confirm("The entry will be permanently deleted. Are you sure?")){
-        API.deleteEntry(entryID)
+    // DELETE entry
+    if (idArray[0] == "delete") {
+        // Pop Up Confirm
+        if (confirm("The entry will be permanently deleted. Are you sure?")) {
+            API.deleteEntry(entryID)
+                .then(data => {
+                    API.get()
+                        .then(data => renderJournalEntries(data))
+                })
+        }
+    }
+    // EDIT entry
+    else if (idArray[0] == "edit") {
+        API.get(entryID)
         .then(data => {
-            API.get()
-            .then(data => renderJournalEntries(data))
+            updateFormFields(data)
         })
     }
-
 })
 
+// Update Form Fields
+const updateFormFields = (data = null) => {
+    let date = document.querySelector("#dateForm");
+    let concepts = document.querySelector("#cenceptsForm");
+    let content = document.querySelector("#contentForm");
+    let mood = document.querySelector("#moodForm");
+    let idEdit = document.querySelector("#recipeId");
+    if (!!data) {
+        date.value = data["date"]
+        concepts.value = data["concepts"]
+        content.value = data["content"]
+        mood.value = data["mood"]
+        idEdit.value = data["id"]
+    }
+}
 
 // Render the Database once the Document is fully loaded 
 document.addEventListener("DOMContentLoaded", () => {
